requirements.txt
colorama==0.4.6
Levenshtein==0.20.9
MouseInfo==0.1.3
mysql-connector-python==8.0.30
Pillow==9.4.0
protobuf==3.20.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypng==0.20220715.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyRect==0.2.0
PyScreeze==0.1.28
python-Levenshtein==0.20.9
pytweening==1.0.4
pywin32==305
PyYAML==6.0
qrcode==7.4.2
rapidfuzz==2.13.7
typing_extensions==4.5.0
SQLAlchemy-Utils
SQLALchemy==2.0.20